requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.1.18
chardet==3.0.4
click==6.7
cssselect==1.0.3
feedfinder2==0.0.4
feedparser==5.2.1
Flask==0.12.2
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
jieba3k==0.35.1
Jinja2==2.10
lxml==4.1.1
MarkupSafe==1.0
newspaper3k==0.2.6
nltk==3.2.5
Pillow==5.0.0
python-dateutil==2.6.1
PyYAML==3.12
readability==0.2
readability-lxml==0.6.2
requests==2.18.4
requests-file==1.4.3
six==1.11.0
tldextract==2.2.0
urllib3==1.22
Werkzeug==0.14.1
WTForms==2.1
